+++ src/interaction/helper.js
@@ -18,7 +18,7 @@ module.exports = {
     }
     return colDef;
   },
-  _getFieldRange(scale, limitRange, type) {
+  getFieldRange(scale, limitRange, type) {
     if (!scale) return [ 0, 1 ];
     let minRatio = 0;
     let maxRatio = 0;
@@ -36,7 +36,7 @@ module.exports = {
     }
     return [ minRatio, maxRatio ];
   },
-  _getLimitRange(data, scale) {
+  getLimitRange(data, scale) {
     let result;
     const { field, type } = scale;
     const values = Util.Array.values(data, field);
@@ -60,5 +60,61 @@ module.exports = {
       result = values;
     }
     return result;
+  },
+  updateCatScale(chart, field, newValues, ticks, values, minIndex, maxIndex) {
+    const colDef = this.getColDef(chart, field);
+    chart.scale(field, Util.mix({}, colDef, {
+      values: newValues,
+      ticks,
+      scale(value) {
+        if (this.type === 'timeCat') {
+          value = this._toTimeStamp(value);
+        }
+        const rangeMin = this.rangeMin();
+        const rangeMax = this.rangeMax();
+        const range = rangeMax - rangeMin;
+        let min;
+        let max;
+        let percent;
+        const currentIndex = values.indexOf(value);
+        if (currentIndex < minIndex) {
+          min = (rangeMin - range) * 5;
+          max = rangeMin;
+          percent = currentIndex / minIndex;
+        } else if (currentIndex > maxIndex) {
+          min = rangeMax;
+          max = (rangeMax + range) * 5;
+          percent = (currentIndex - maxIndex - 1) / (values.length - 1 - maxIndex);
+        } else {
+          const index = this.translate(value);
+          percent = (index) / (this.values.length - 1);
+          min = rangeMin;
+          max = rangeMax;
+        }
+        return min + percent * (max - min);
+      },
+      getTicks() {
+        const self = this;
+        const ticks = this.ticks;
+        const rst = [];
+        Util.each(ticks, tick => {
+          let obj;
+          if (Util.isObject(tick)) {
+            obj = tick;
+          } else {
+            let value = self.scale(tick);
+            value = value >= 0 && value <= 1 ? value : NaN;
+
+            obj = {
+              text: Util.isString(tick) ? tick : self.getText(tick),
+              value,
+              tickValue: tick // 用于坐标轴上文本动画时确定前后帧的对应关系
+            };
+          }
+          rst.push(obj);
+        });
+        return rst;
+      }
+    }));
   }
 };
